+++ bin/fix-lint-backend.sh
@@ -10,9 +10,9 @@ bin/sortignore.sh
 ####################
 ###### Python ######
 ###################
-poetry run ruff check --fix .
-poetry run ruff format .
-# poetry run djlint templates --profile=django --reformat
+uv run ruff check --fix .
+uv run ruff format .
+# uv run djlint templates --profile=django --reformat
 
 ############################################
 ##### Javascript, JSON, Markdown, YAML #####