+++ picopt/walk.py
@@ -1,6 +1,5 @@
 """Walk the directory trees and files and call the optimizers."""
 
-import os
 import shutil
 import traceback
 from multiprocessing.pool import ApplyResult, Pool
@@ -165,9 +164,9 @@ class Walk:
 
         results = []
         files = []
-        dir_path: Path = path_info.path  # type: ignore
+        dir_path: Path = path_info.path  # type: ignore[reportAssignmentType]
 
-        for name in sorted(os.listdir(dir_path)):
+        for name in sorted(dir_path.iterdir()):
             entry_path = dir_path / name
             if entry_path.is_dir():
                 path_info = PathInfo(
@@ -334,7 +333,13 @@ class Walk:
         for top_path in self._top_paths:
             dirpath = Treestamps.get_dir(top_path)
             is_case_sensitive = self._is_case_sensitive(dirpath)
-            path_info = PathInfo(dirpath, 0.0, True, is_case_sensitive, path=top_path)
+            path_info = PathInfo(
+                dirpath,
+                0.0,
+                convert=True,
+                is_case_sensitive=is_case_sensitive,
+                path=top_path,
+            )
             result = self.walk_file(path_info)
             if not result:
                 continue