+++ tests/integration/test_images_dir.py
@@ -10,10 +10,9 @@ from tests.integration.base_test_images import BaseTestImagesDir
 __all__ = ()
 FNS = {
     "07themecamplist.pdf": (93676, 93676, ("pdf", 93676), ("pdf", 93676)),
-    "test_animated_gif.gif": (16383, 16358, ("png", 13389), ("webp", 11894)),
-    "test_animated_png.png": (63435, 63435, ("png", 63435), ("webp", 53430)),
-    "test_animated_webp.webp": (13610, 12178, ("webp", 12178), ("webp", 12178)),
-    "test_png.png": (7967, 4149, ("png", 4149), ("webp", 3870)),
+    "test_animated_gif.gif": (16383, 16358, ("png", 24255), ("webp", 11892)),
+    "test_animated_webp.webp": (13610, 12174, ("webp", 12174), ("webp", 12174)),
+    "test_png.png": (7967, 4150, ("png", 4150), ("webp", 3870)),
     "test_pre-optimized_jpg.jpg": (
         22664,
         22664,
@@ -29,17 +28,32 @@ FNS = {
         ("webp", 3798),
         ("webp", 3798),
     ),
-    "mri.tif": (230578, 230578, ("png", 131743), ("webp", 114740)),
+    "mri.tif": (230578, 230578, ("png", 129479), ("webp", 116954)),
     "test_webp_lossy.webp": (2764, 2764, ("webp", 2764), ("webp", 2764)),
-    "test_bmp.bmp": (141430, 141430, ("png", 67236), ("webp", 47524)),
     "test_png_16rgba.png": (3435, 2097, ("png", 2097), ("webp", 1142)),
-    "test_pnm.pnm": (27661, 27661, ("png", 15510), ("webp", 12808)),
     "test_jpg.jpg": (97373, 87913, ("jpg", 87913), ("jpg", 87913)),
 }
 if platform.system() == "Darwin":
     FNS.update(
         {
-            "test_gif.gif": (138952, 138944, ("png", 112137), ("webp", 107924)),
+            "test_animated_png.png": (63435, 63058, ("png", 63058), ("webp", 53388)),
+            "test_gif.gif": (138952, 138944, ("png", 112290), ("webp", 108058)),
+            "test_pre-optimized_png.png": (
+                256572,
+                256572,
+                ("png", 256572),
+                ("webp", 197088),
+            ),
+            "eight.tif": (59640, 59640, ("png", 30585), ("webp", 25012)),
+            "test_bmp.bmp": (141430, 141430, ("png", 67236), ("webp", 47436)),
+            "test_pnm.pnm": (27661, 27661, ("png", 15510), ("webp", 12758)),
+        }
+    )
+else:
+    FNS.update(
+        {
+            "test_animated_png.png": (63435, 63058, ("png", 63058), ("webp", 53422)),
+            "test_gif.gif": (138952, 138944, ("png", 112290), ("webp", 107924)),
             "test_pre-optimized_png.png": (
                 256572,
                 256572,
@@ -47,19 +61,8 @@ if platform.system() == "Darwin":
                 ("webp", 197726),
             ),
             "eight.tif": (59640, 59640, ("png", 30585), ("webp", 24974)),
-        }
-    )
-else:
-    FNS.update(
-        {
-            "test_gif.gif": (138952, 138944, ("png", 112137), ("webp", 107952)),
-            "test_pre-optimized_png.png": (
-                256572,
-                256572,
-                ("png", 256572),
-                ("webp", 197680),
-            ),
-            "eight.tif": (59640, 59640, ("png", 30585), ("webp", 24982)),
+            "test_bmp.bmp": (141430, 141430, ("png", 67236), ("webp", 47524)),
+            "test_pnm.pnm": (27661, 27661, ("png", 15510), ("webp", 12808)),
         }
     )
 
@@ -95,7 +98,7 @@ class TestImagesDir(BaseTestImagesDir):
         """Test convert to PNG."""
         args = (
             PROGRAM_NAME,
-            "-rvvx",
+            "-rvvbx",
             "BMP,PPM,SVG,TIFF",
             "-c",
             "PNG",