+++ NEWS.md
@@ -1,5 +1,13 @@
 # 📰 Picopt News
 
+## v4.1.0
+
+- -T --test option becomes -d --dry-run
+- Fix tif & animated gifs not converted to pngs when asked.
+- animated pngs better optimized with oxipng
+- updated treestamps
+- uv build system
+
 ## v4.0.4
 
 - Pillow 11 support.